Congregated additional functionality.
* Removed some old and unused functionality.
* Combined related code into appropriate modules.
