+++ MANIFEST.in
@@ -11,3 +11,4 @@ recursive-exclude * *.py[co]
 recursive-include docs *.rst conf.py Makefile make.bat *.jpg *.png *.gif
 
 graft conf
+graft data
